+++ src/CHANGELOG.rst
@@ -13,6 +13,7 @@ v4.0dev25 = New Dev Release (???)
 - Added GoldyBot "send_modal()" method in commands utils.
 - Added "Context" to main GoldyBot module.
 - Fixed bug with nested sub commands.
+- Added default timezone setting in Timestamps.
 
 v4.0dev24 = New Dev Release (01/10/2022)
 ---------------------------