+++ src/CHANGELOG.rst
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+v4.1dev1 = Full Under the Hood Rewrite (???)
+---------------------------
+
+
 v4.0dev25 = New Dev Release (???)
 ---------------------------
 - Fixed /cache command.